@@ -42,14 +42,13 @@ setup(
     classifiers=[
         'Development Status :: 5 - Production/Stable',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
         'Programming Language :: Python :: 3.11',
         'Programming Language :: Python :: 3.12',
         'Programming Language :: Python :: 3.13',
         'Programming Language :: Python :: 3.14',
     ],
-    python_requires='>=3.9',
+    python_requires='>=3.10',
     install_requires=[
         f'large-image{limit_version}',
         'wsidicom>=0.9.0,!=0.21.3',
