@@ -17,8 +17,8 @@
 import pickle
 import threading
 import time
-from collections.abc import Iterable, Sized
-from typing import Any, Callable, Optional, TypeVar, Union, cast
+from collections.abc import Callable, Iterable, Sized
+from typing import Any, Optional, TypeVar, cast
 
 from typing_extensions import Buffer
 
@@ -32,9 +32,9 @@ class RedisCache(BaseCache):
     """Use redis as the backing cache."""
 
     def __init__(
-            self, url: Union[str, list[str]] = '127.0.0.1:6379',
-            username: Optional[str] = None, password: Optional[str] = None,
-            getsizeof: Optional[Callable[[_VT], float]] = None,
+            self, url: str | list[str] = '127.0.0.1:6379',
+            username: str | None = None, password: str | None = None,
+            getsizeof: Callable[[_VT], float] | None = None,
             mustBeAvailable: bool = False) -> None:
         import redis
         from redis.client import Redis
